provide an indirection for loading the JAXB implementation
