+++ .travis.yml
@@ -1,5 +1,5 @@
 language: java
-env: VERSION=2.8.5-SNAPSHOT
+env: VERSION=2.9.0-SNAPSHOT
 matrix:
   include:
     - dist: trusty
