+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## XMLUnit for Java 2.8.5 - /no released, yet/
+## XMLUnit for Java 2.9.0 - /no released, yet/
 
 ## XMLUnit for Java 2.8.4 - /Released 2021-12-16/
