+++ xmlunit-assertj/pom.xml
@@ -18,7 +18,7 @@
   <parent>
     <artifactId>xmlunit-parent</artifactId>
     <groupId>org.xmlunit</groupId>
-    <version>2.8.5-SNAPSHOT</version>
+    <version>2.9.0-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
