+++ xmlunit-core/src/main/java/org/xmlunit/jaxb/DefaultJaxbBuilderFactory.java
@@ -1,0 +1,29 @@
+/*
+  This file is licensed to You under the Apache License, Version 2.0
+  (the "License"); you may not use this file except in compliance with
+  the License.  You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+ */
+
+package org.xmlunit.jaxb;
+
+import org.xmlunit.builder.JaxbBuilder;
+import org.xmlunit.builder.JaxbBuilderFactory;
+
+/**
+ * Default implementation that uses {@code javax.xml.bind}.
+ * @since 2.9.0
+ */
+public class DefaultJaxbBuilderFactory implements JaxbBuilderFactory {
+    @Override
+    public JaxbBuilder create(Object object) {
+        return new JavaxJaxbBuilder(object);
+    }
+}
