+++ xmlunit-core/src/test/java/org/xmlunit/jaxb/JavaxJaxbBuilderTest.java
@@ -12,7 +12,7 @@
   limitations under the License.
  */
 
-package org.xmlunit.builder;
+package org.xmlunit.jaxb;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.CoreMatchers.containsString;
@@ -39,7 +39,7 @@ import javax.xml.transform.stream.StreamResult;
 import java.io.StringWriter;
 
 
-public class JaxbBuilderTest {
+public class JavaxJaxbBuilderTest {
 
     @Test
     public void testMarchal_withJaxbRootObject_shouldReturnSource() throws Exception {
@@ -47,7 +47,7 @@ public class JaxbBuilderTest {
         final Object testObject = createRootNode("123");
         
         // run test
-        final Source saxSource = new JaxbBuilder(testObject).build();
+        final Source saxSource = new JavaxJaxbBuilder(testObject).build();
 
         // validate result
         final String xmlString = toString(saxSource);
@@ -67,7 +67,7 @@ public class JaxbBuilderTest {
         Marshaller marshaller = JAXBContext.newInstance(testObject.getClass()).createMarshaller();
 
         // run test
-        final Source saxSource = new JaxbBuilder(testObject).withMarshaller(marshaller ).build();
+        final Source saxSource = new JavaxJaxbBuilder(testObject).withMarshaller(marshaller ).build();
 
         // validate result
         final String xmlString = toString(saxSource);
@@ -86,7 +86,7 @@ public class JaxbBuilderTest {
         final Object testObject = createComplexNode("123");
         
         // run test
-        final Source saxSource = new JaxbBuilder(testObject).build();
+        final Source saxSource = new JavaxJaxbBuilder(testObject).build();
 
         // validate result
         final String xmlString = toString(saxSource);
@@ -101,7 +101,7 @@ public class JaxbBuilderTest {
         final Object testObject = createComplexNode("123");
 
         // run test
-        final Source saxSource = new JaxbBuilder(testObject).useObjectFactory().build();
+        final Source saxSource = new JavaxJaxbBuilder(testObject).useObjectFactory().build();
 
         // validate result
         final String xmlString = toString(saxSource);
@@ -121,7 +121,7 @@ public class JaxbBuilderTest {
         final Object testJAXBElement = new JAXBElement<ComplexNode>(name, ComplexNode.class, createComplexNode("123"));
 
         // run test
-        final Source saxSource = new JaxbBuilder(testJAXBElement).build();
+        final Source saxSource = new JavaxJaxbBuilder(testJAXBElement).build();
 
         // validate result
         final String xmlString = toString(saxSource);
