+++ xmlunit-legacy/pom.xml
@@ -18,7 +18,7 @@
   <parent>
     <groupId>org.xmlunit</groupId>
     <artifactId>xmlunit-parent</artifactId>
-    <version>2.8.5-SNAPSHOT</version>
+    <version>2.9.0-SNAPSHOT</version>
   </parent>
 
   <groupId>org.xmlunit</groupId>
